Remove redundant dependencies
Most of the removed dependencies are implicit, the other packages depend
on them and they will be instaled even if not listed explicitly.

openshift is opposite. Nothing depends on it, however drupal depends on
kubernetes which is also an openshift's dependency. We can cut out the
middle man and specify kubernetes directly.