ZFS package installation fails
As of the date of this commit, ZFS package cannot be installed to the
Raspberry PI OS from the standard image. The build script for zfs-dkms
package supports kernels between 3.10 and 5.10 while the package itself
comes with kernel 5.15.32-v8+.

apt update; apt upgrade is not good enough. The only way to get a newer
zfs-dkms package is to use a backport repository.